requirements.txt
Django==2.1.1
djangorestframework==3.9
django-filter
django-crispy-forms
django-debug-toolbar
psycopg2-binary
gunicorn
django-heroku
markdown
drf-yasg
flex
swagger_spec_validator
